pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-clickhouse-backend = "^1.0.2"
django = "^4.2.2"
notalib = "^2.1.0"


[tool.poetry.group.dev.dependencies]
faker = "^18.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
